Refactor: Handle delete in child class
